refactor(dap): split debug_adapter.rs into domain modules (#1639)
Extract four domain modules from the monolithic debug_adapter/mod.rs:

- dispatch.rs: request handling and dispatch (handle_request, dispatch_request)
- transport.rs: I/O transport layer (run, run_socket, run_with_io)
- parsing.rs: debugger output parsing (normalize, infer types, parse frames/variables)
- safe_eval.rs: expression safety validation (validate_safe_expression and helpers)

mod.rs retains core types, regex patterns, constructors, utility methods, and
integration tests. All 96 tests pass with no behavioral changes.
