+++ crates/perl-dap/src/debug_adapter/mod.rs
@@ -11,6 +11,11 @@ mod output;
 mod process;
 mod variables;
 
+mod dispatch;
+mod parsing;
+pub(crate) mod safe_eval;
+mod transport;
+
 use crate::feature_catalog::has_feature as catalog_has_feature;
 use crate::inline_values::collect_inline_values;
 use crate::protocol::{
@@ -59,6 +64,7 @@ use nix::sys::signal::{self, Signal};
 #[cfg(unix)]
 use nix::unistd::Pid;
 use regex::Regex;
+use safe_eval::validate_safe_expression;
 
 /// Poison-safe mutex lock that recovers from poisoned state
 fn lock_or_recover<'a, T>(mutex: &'a Mutex<T>, ctx: &'static str) -> MutexGuard<'a, T> {
@@ -564,245 +570,6 @@ impl DebugAdapter {
         }
     }
 
-    /// Run the debug adapter server
-    pub fn run(&mut self) -> io::Result<()> {
-        self.run_with_io(io::stdin(), io::stdout())
-    }
-
-    /// Run the debug adapter over a TCP socket transport.
-    ///
-    /// This binds to `127.0.0.1:<port>`, accepts one client connection, and
-    /// serves the DAP session on that stream.
-    pub fn run_socket(&mut self, port: u16) -> io::Result<()> {
-        let listener = TcpListener::bind(("127.0.0.1", port))?;
-        eprintln!("DAP socket transport listening on 127.0.0.1:{port}");
-
-        let (stream, peer_addr) = listener.accept()?;
-        eprintln!("DAP socket client connected: {peer_addr}");
-
-        let reader_stream = stream.try_clone()?;
-        self.run_with_io(reader_stream, stream)
-    }
-
-    /// Shared DAP transport loop used by stdio and socket modes.
-    fn run_with_io<R, W>(&mut self, input: R, output: W) -> io::Result<()>
-    where
-        R: Read,
-        W: Write + Send + 'static,
-    {
-        // Create a shared writer to prevent interleaving between the main loop
-        // and the event handler thread.
-        let shared_writer: Arc<Mutex<W>> = Arc::new(Mutex::new(output));
-        let event_writer = Arc::clone(&shared_writer);
-
-        // Create channel for asynchronous events.
-        let (tx, rx) = channel::<DapMessage>();
-        self.event_sender = Some(tx.clone());
-
-        thread::spawn(move || {
-            while let Ok(msg) = rx.recv() {
-                let framed = match serde_json::to_vec(&msg) {
-                    Ok(payload) => frame(&payload),
-                    Err(e) => {
-                        eprintln!("Failed to serialize DAP message: {} - {:#?}", e, msg);
-                        continue;
-                    }
-                };
-
-                let mut writer = lock_or_recover(&event_writer, "event_writer");
-                if let Err(e) = writer.write_all(&framed) {
-                    eprintln!("Failed to write DAP frame in event handler: {}", e);
-                    continue;
-                }
-                if let Err(e) = writer.flush() {
-                    eprintln!("Failed to flush DAP frame in event handler: {}", e);
-                }
-            }
-            eprintln!("Event handler thread terminating - channel closed");
-        });
-
-        let mut reader = BufReader::new(input);
-        let mut framer = ContentLengthFramer::new();
-        let mut read_buf = [0u8; 8 * 1024];
-
-        loop {
-            let bytes_read = reader.read(&mut read_buf)?;
-            if bytes_read == 0 {
-                return Ok(());
-            }
-
-            framer.push(&read_buf[..bytes_read]);
-
-            loop {
-                let body = match framer.try_next() {
-                    Ok(Some(body)) => body,
-                    Ok(None) => break,
-                    Err(error) => {
-                        eprintln!("Failed to parse DAP transport frame: {error}");
-                        continue;
-                    }
-                };
-
-                let msg = match serde_json::from_slice::<DapMessage>(&body) {
-                    Ok(msg) => msg,
-                    Err(_) => {
-                        eprintln!(
-                            "Failed to parse DAP message: {}",
-                            String::from_utf8_lossy(&body)
-                        );
-                        continue;
-                    }
-                };
-
-                let DapMessage::Request { seq, command, arguments } = msg else {
-                    continue;
-                };
-
-                let response = self.dispatch_request(seq, &command, arguments);
-                let payload = match serde_json::to_vec(&response) {
-                    Ok(payload) => payload,
-                    Err(e) => {
-                        eprintln!("Failed to serialize DAP response: {}", e);
-                        continue;
-                    }
-                };
-
-                let framed = frame(&payload);
-                let mut writer = lock_or_recover(&shared_writer, "response_writer");
-                writer.write_all(&framed)?;
-                writer.flush()?;
-
-                // DAP requires this event only after initialize response is sent.
-                if command == "initialize"
-                    && Self::response_succeeded_for_command(&response, "initialize")
-                {
-                    self.send_event("initialized", None);
-                }
-            }
-        }
-    }
-
-    /// Handle a DAP request
-    pub fn handle_request(
-        &mut self,
-        request_seq: i64,
-        command: &str,
-        arguments: Option<Value>,
-    ) -> DapMessage {
-        eprintln!("DAP request: {} {:?}", command, arguments);
-
-        let response = self.dispatch_request(request_seq, command, arguments);
-
-        // Preserve existing direct-call behavior for tests and in-memory usage.
-        if command == "initialize" && Self::response_succeeded_for_command(&response, "initialize")
-        {
-            self.send_event("initialized", None);
-        }
-
-        response
-    }
-
-    /// Handle a DAP request (mock version for testing)
-    pub fn handle_request_mock(
-        &mut self,
-        request_seq: i64,
-        command: &str,
-        arguments: Option<Value>,
-    ) -> DapMessage {
-        eprintln!("DAP request (mock): {} {:?}", command, arguments);
-
-        if command == "attach" {
-            let seq = self.next_seq();
-            return DapMessage::Response {
-                seq,
-                request_seq,
-                success: false,
-                command: "attach".to_string(),
-                body: None,
-                message: Some("Attach not yet fully implemented".to_string()),
-            };
-        }
-
-        let response = self.dispatch_request(request_seq, command, arguments);
-        if command == "initialize" && Self::response_succeeded_for_command(&response, "initialize")
-        {
-            self.send_event("initialized", None);
-        }
-        response
-    }
-
-    fn dispatch_request(
-        &mut self,
-        request_seq: i64,
-        command: &str,
-        arguments: Option<Value>,
-    ) -> DapMessage {
-        let seq = self.next_seq();
-
-        match command {
-            "initialize" => self.handle_initialize(seq, request_seq, arguments),
-            "launch" => self.handle_launch(seq, request_seq, arguments),
-            "attach" => self.handle_attach(seq, request_seq, arguments),
-            "disconnect" => self.handle_disconnect(seq, request_seq, arguments),
-            "terminate" => self.handle_terminate(seq, request_seq, arguments),
-            "setBreakpoints" => self.handle_set_breakpoints(seq, request_seq, arguments),
-            "setFunctionBreakpoints" => {
-                self.handle_set_function_breakpoints(seq, request_seq, arguments)
-            }
-            "setExceptionBreakpoints" => {
-                self.handle_set_exception_breakpoints(seq, request_seq, arguments)
-            }
-            "configurationDone" => self.handle_configuration_done(seq, request_seq),
-            "threads" => self.handle_threads(seq, request_seq),
-            "stackTrace" => self.handle_stack_trace(seq, request_seq, arguments),
-            "scopes" => self.handle_scopes(seq, request_seq, arguments),
-            "variables" => self.handle_variables(seq, request_seq, arguments),
-            "setVariable" => self.handle_set_variable(seq, request_seq, arguments),
-            "continue" => self.handle_continue(seq, request_seq, arguments),
-            "next" => self.handle_next(seq, request_seq, arguments),
-            "stepIn" => self.handle_step_in(seq, request_seq, arguments),
-            "stepOut" => self.handle_step_out(seq, request_seq, arguments),
-            "pause" => self.handle_pause(seq, request_seq, arguments),
-            "evaluate" => self.handle_evaluate(seq, request_seq, arguments),
-            "inlineValues" => self.handle_inline_values(seq, request_seq, arguments),
-            "breakpointLocations" => self.handle_breakpoint_locations(seq, request_seq, arguments),
-            "source" => self.handle_source(seq, request_seq, arguments),
-            "loadedSources" => self.handle_loaded_sources(seq, request_seq, arguments),
-            "modules" => self.handle_modules(seq, request_seq, arguments),
-            "completions" => self.handle_completions(seq, request_seq, arguments),
-            "exceptionInfo" => self.handle_exception_info(seq, request_seq, arguments),
-            "restart" => self.handle_restart(seq, request_seq, arguments),
-            "setExpression" => self.handle_set_expression(seq, request_seq, arguments),
-            "dataBreakpointInfo" => self.handle_data_breakpoint_info(seq, request_seq, arguments),
-            "setDataBreakpoints" => self.handle_set_data_breakpoints(seq, request_seq, arguments),
-            "cancel" => self.handle_cancel(seq, request_seq, arguments),
-            "stepInTargets" => self.handle_step_in_targets(seq, request_seq, arguments),
-            "gotoTargets" => self.handle_goto_targets(seq, request_seq, arguments),
-            "goto" => self.handle_goto(seq, request_seq, arguments),
-            "restartFrame" => self.handle_restart_frame(seq, request_seq, arguments),
-            "terminateThreads" => self.handle_terminate_threads(seq, request_seq, arguments),
-            _ => DapMessage::Response {
-                seq,
-                request_seq,
-                success: false,
-                command: command.to_string(),
-                body: None,
-                message: Some(format!("Unknown command: {}", command)),
-            },
-        }
-    }
-
-    fn response_succeeded_for_command(response: &DapMessage, expected_command: &str) -> bool {
-        matches!(
-            response,
-            DapMessage::Response {
-                success: true,
-                command,
-                ..
-            } if command == expected_command
-        )
-    }
-
     /// Get next sequence number (monotonically increasing, poison-safe)
     fn next_seq(&self) -> i64 {
         let mut seq = lock_or_recover(&self.seq, "next_seq");
@@ -828,43 +595,6 @@ impl DebugAdapter {
     /// Allocate a unique marker id used for framed debugger output capture.
     fn next_debugger_marker_id(&self) -> u64 {
         self.debugger_output_marker.fetch_add(1, Ordering::Relaxed)
-    }
-
-    /// Normalize debugger output lines for deterministic parsing by:
-    /// - removing ANSI escape sequences
-    /// - stripping debugger prompt prefixes (e.g. `DB<1>`)
-    fn normalize_debugger_output_line(line: &str) -> String {
-        let mut normalized = if let Some(re) = ansi_escape_re() {
-            re.replace_all(line, "").into_owned()
-        } else {
-            line.to_string()
-        };
-
-        if let Some(prompt_start) = normalized.find("DB<")
-            && let Some(prompt_end) = normalized[prompt_start..].find('>')
-        {
-            let content_start = prompt_start + prompt_end + 1;
-            normalized = normalized[content_start..].to_string();
-        }
-
-        normalized.trim().to_string()
-    }
-
-    /// Infer a coarse DAP value type from literal-like debugger output.
-    fn infer_debugger_value_type(text: &str) -> String {
-        if text == "undef" {
-            "undef".to_string()
-        } else if text.parse::<i64>().is_ok() {
-            "integer".to_string()
-        } else if text.parse::<f64>().is_ok() {
-            "number".to_string()
-        } else if text.starts_with('[') && text.ends_with(']') {
-            "array".to_string()
-        } else if text.starts_with('{') && text.ends_with('}') {
-            "hash".to_string()
-        } else {
-            "string".to_string()
-        }
     }
 
     /// Write a debugger command and flush immediately so output framing remains ordered.
@@ -978,601 +708,7 @@ impl DebugAdapter {
             }
         }
     }
-
-    /// Convert microcrate rendered variables into adapter-local protocol values.
-    fn rendered_to_variable(rendered: RenderedVariable) -> Variable {
-        Variable {
-            name: rendered.name,
-            value: rendered.value,
-            type_: rendered.type_name,
-            variables_reference: Self::i64_to_i32_saturating(rendered.variables_reference),
-            named_variables: rendered.named_variables.map(Self::i64_to_i32_saturating),
-            indexed_variables: rendered.indexed_variables.map(Self::i64_to_i32_saturating),
-        }
-    }
-
-    /// Determine if a variable name should appear in a given scope.
-    fn scope_allows_variable_name(scope_type: i32, name: &str) -> bool {
-        match scope_type {
-            // Locals
-            1 => !name.contains("::"),
-            // Package variables (qualified)
-            2 => name.contains("::"),
-            // Globals/specials
-            3 => {
-                matches!(name, "$_" | "@ARGV" | "%ENV" | "$!" | "$@" | "$/" | "$|" | "$0" | "$^W")
-                    || name.starts_with("$^")
-            }
-            _ => true,
-        }
-    }
-
-    /// Convert parsed stack frames from `perl-dap-stack` into local DAP response frames.
-    fn parse_stack_frames_from_text(output: &str) -> Vec<StackFrame> {
-        let mut parser = PerlStackParser::new();
-        parser
-            .parse_stack_trace(output)
-            .into_iter()
-            .map(|frame| {
-                let source = frame.source.unwrap_or_default();
-                let path = source.path.unwrap_or_else(|| "<unknown>".to_string());
-                let name = source.name.or_else(|| {
-                    std::path::Path::new(&path)
-                        .file_name()
-                        .and_then(|n| n.to_str())
-                        .map(ToString::to_string)
-                });
-                StackFrame {
-                    id: Self::i64_to_i32_saturating(frame.id),
-                    name: frame.name,
-                    source: Source { name, path, source_reference: None },
-                    line: Self::i64_to_i32_saturating(frame.line),
-                    column: Self::i64_to_i32_saturating(frame.column),
-                    end_line: frame.end_line.map(Self::i64_to_i32_saturating),
-                    end_column: frame.end_column.map(Self::i64_to_i32_saturating),
-                }
-            })
-            .collect()
-    }
-
-    /// Filter out internal debugger and shim frames from user-visible stack traces.
-    fn filter_user_visible_frames(frames: Vec<StackFrame>) -> Vec<StackFrame> {
-        frames
-            .into_iter()
-            .filter(|frame| {
-                !is_internal_frame_name_and_path(&frame.name, Some(frame.source.path.as_str()))
-            })
-            .collect()
-    }
-
-    /// Parse variables from debugger output lines using microcrate parser/renderer.
-    fn parse_scope_variables_from_lines(
-        lines: &[String],
-        variables_ref: i32,
-        start: usize,
-        count: usize,
-    ) -> (Vec<Variable>, HashMap<i32, Vec<Variable>>) {
-        let parser = VariableParser::new();
-        let renderer = PerlVariableRenderer::new();
-        let scope_type = variables_ref % 10;
-        let mut seen = HashSet::new();
-        let mut parsed = Vec::new();
-
-        for line in lines.iter().rev() {
-            let normalized = Self::normalize_debugger_output_line(line);
-            let text = normalized.trim();
-            if text.is_empty() {
-                continue;
-            }
-            if let Ok((name, value)) = parser.parse_assignment(text) {
-                if !Self::scope_allows_variable_name(scope_type, &name) {
-                    continue;
-                }
-                if seen.insert(name.clone()) {
-                    parsed.push((name, value));
-                }
-                if parsed.len() >= 256 {
-                    break;
-                }
-            }
-        }
-
-        parsed.reverse();
-        parsed.sort_unstable_by(|(left, _), (right, _)| left.cmp(right));
-
-        let mut top_level = Vec::new();
-        let mut child_cache = HashMap::new();
-        for (idx, (name, value)) in parsed.into_iter().skip(start).take(count).enumerate() {
-            let child_ref = variables_ref.saturating_mul(1000).saturating_add(
-                Self::i64_to_i32_saturating(i64::try_from(idx + 1).unwrap_or(i64::from(i32::MAX))),
-            );
-            let rendered = if value.is_expandable() {
-                renderer.render_with_reference(&name, &value, i64::from(child_ref))
-            } else {
-                renderer.render(&name, &value)
-            };
-            top_level.push(Self::rendered_to_variable(rendered));
-
-            if value.is_expandable() {
-                let children = renderer
-                    .render_children(&value, 0, 256)
-                    .into_iter()
-                    .map(Self::rendered_to_variable)
-                    .collect::<Vec<_>>();
-                if !children.is_empty() {
-                    child_cache.insert(child_ref, children);
-                }
-            }
-        }
-
-        (top_level, child_cache)
-    }
-
-    /// Parse variables from recent debugger output using microcrate parser/renderer.
-    fn parse_scope_variables_from_output(
-        &self,
-        variables_ref: i32,
-        start: usize,
-        count: usize,
-    ) -> (Vec<Variable>, HashMap<i32, Vec<Variable>>) {
-        let lines = self.snapshot_recent_output_lines();
-        Self::parse_scope_variables_from_lines(&lines, variables_ref, start, count)
-    }
-
-    /// Parse evaluate output from debugger lines into a DAP result payload.
-    fn parse_evaluate_result_from_lines(
-        lines: &[String],
-        expression: &str,
-        allow_fallback_line: bool,
-    ) -> Option<(String, String)> {
-        if lines.is_empty() {
-            return None;
-        }
-
-        let parser = VariableParser::new();
-        let renderer = PerlVariableRenderer::new();
-
-        for line in lines.iter().rev() {
-            let normalized = Self::normalize_debugger_output_line(line);
-            let text = normalized.trim();
-            if text.is_empty() || prompt_re().is_some_and(|re| re.is_match(text)) {
-                continue;
-            }
-
-            if let Ok((name, value)) = parser.parse_assignment(text) {
-                let rendered = renderer.render(&name, &value);
-                let type_name = rendered.type_name.unwrap_or_else(|| "string".to_string());
-                // Prefer direct matches for the evaluated expression, but allow fallback assignment.
-                if name == expression || text.starts_with(expression) || text.contains(expression) {
-                    return Some((rendered.value, type_name));
-                }
-                if !allow_fallback_line {
-                    continue;
-                }
-                return Some((rendered.value, type_name));
-            }
-
-            if allow_fallback_line {
-                return Some((text.to_string(), Self::infer_debugger_value_type(text)));
-            }
-        }
-
-        None
-    }
-
-    /// Parse evaluate output from recent debugger lines into a DAP result payload.
-    fn parse_evaluate_result_from_output(&self, expression: &str) -> Option<(String, String)> {
-        let lines = self.snapshot_recent_output_lines();
-        Self::parse_evaluate_result_from_lines(&lines, expression, true)
-    }
-
-    /// Build deterministic placeholder variables used when debugger output is unavailable.
-    fn fallback_scope_variables(variables_ref: i32, start: usize, count: usize) -> Vec<Variable> {
-        let variables = match variables_ref % 10 {
-            1 => vec![
-                Variable {
-                    name: "$self".to_string(),
-                    value: "blessed(My::Module)".to_string(),
-                    type_: Some("hash".to_string()),
-                    variables_reference: variables_ref.saturating_mul(100) + 2,
-                    named_variables: Some(5),
-                    indexed_variables: None,
-                },
-                Variable {
-                    name: "@_".to_string(),
-                    value: "array(size=0)".to_string(),
-                    type_: Some("array".to_string()),
-                    variables_reference: variables_ref.saturating_mul(100) + 1,
-                    named_variables: None,
-                    indexed_variables: Some(0),
-                },
-            ],
-            2 => vec![Variable {
-                name: "$VERSION".to_string(),
-                value: "\"1.0.0\"".to_string(),
-                type_: Some("scalar".to_string()),
-                variables_reference: 0,
-                named_variables: None,
-                indexed_variables: None,
-            }],
-            3 => vec![Variable {
-                name: "$_".to_string(),
-                value: "undef".to_string(),
-                type_: Some("scalar".to_string()),
-                variables_reference: 0,
-                named_variables: None,
-                indexed_variables: None,
-            }],
-            _ => Vec::new(),
-        };
-
-        variables.into_iter().skip(start).take(count).collect()
-    }
-
-    /// Parse stack trace output from Perl debugger "T" command
-    ///
-    /// AC8.2: Parse caller() + %DB::sub data from Perl debugger
-    ///
-    /// The Perl debugger "T" command outputs stack traces in formats like:
-    /// ```text
-    /// $ = main::compute_sum() called from file /app/main.pl line 15
-    /// $ = main::process_data() called from file /app/main.pl line 10
-    /// ```
-    ///
-    /// Or with frame numbers:
-    /// ```text
-    /// # 0 main::helper at /app/script.pl line 20
-    /// # 1 Foo::bar called at /app/lib/Foo.pm line 15
-    /// # 2 main::start at /app/script.pl line 5
-    /// ```
-    ///
-    /// Returns a vector of StackFrame structs with accurate line numbers,
-    /// source paths, and package-qualified function names.
-    #[cfg(test)]
-    fn parse_stack_trace(output: &str) -> Vec<StackFrame> {
-        let mut frames = Vec::new();
-        let mut frame_id = 1;
-
-        for line in output.lines() {
-            // Try to match stack frame format
-            if let Some(re) = stack_frame_re() {
-                if let Some(caps) = re.captures(line) {
-                    let func = caps.name("func").map(|m| m.as_str()).unwrap_or("main");
-                    let file = caps.name("file").map(|m| m.as_str()).unwrap_or("<unknown>");
-                    let line_num =
-                        caps.name("line").and_then(|m| m.as_str().parse::<i32>().ok()).unwrap_or(1);
-
-                    // Extract file name from path for display
-                    let file_name = file.split(['/', '\\'].as_ref()).next_back().unwrap_or(file);
-
-                    frames.push(StackFrame {
-                        id: frame_id,
-                        name: func.to_string(),
-                        source: Source {
-                            name: Some(file_name.to_string()),
-                            path: file.to_string(),
-                            source_reference: None,
-                        },
-                        line: line_num,
-                        column: 1, // Perl debugger doesn't provide column info by default
-                        end_line: None,
-                        end_column: None,
-                    });
-
-                    frame_id += 1;
-                }
-            }
-        }
-
-        frames
-    }
 }
-
-/// Check if a position in a string is inside single quotes
-/// (conservative: only tracks single-quoted string literals)
-fn is_in_single_quotes(s: &str, idx: usize) -> bool {
-    let mut in_sq = false;
-    let mut escaped = false;
-
-    for (i, ch) in s.char_indices() {
-        if i >= idx {
-            break;
-        }
-        if in_sq {
-            if escaped {
-                escaped = false;
-            } else if ch == '\\' {
-                escaped = true;
-            } else if ch == '\'' {
-                in_sq = false;
-            }
-        } else if ch == '\'' {
-            in_sq = true;
-        }
-    }
-
-    in_sq
-}
-
-/// Check if the match is CORE:: or CORE::GLOBAL:: qualified (must block these)
-fn is_core_qualified(s: &str, op_start: usize) -> bool {
-    let bytes = s.as_bytes();
-
-    // Must have :: immediately before op
-    if op_start < 2 || bytes[op_start - 1] != b':' || bytes[op_start - 2] != b':' {
-        return false;
-    }
-
-    // Extract the identifier right before that ::
-    let end = op_start - 2;
-    let mut start = end;
-    while start > 0 {
-        let b = bytes[start - 1];
-        if b.is_ascii_alphanumeric() || b == b'_' {
-            start -= 1;
-        } else {
-            break;
-        }
-    }
-    let seg = &s[start..end];
-    if seg == "CORE" {
-        return true;
-    }
-    if seg != "GLOBAL" {
-        return false;
-    }
-
-    // If GLOBAL, require CORE::GLOBAL::op
-    if start < 2 || bytes[start - 1] != b':' || bytes[start - 2] != b':' {
-        return false;
-    }
-    let end2 = start - 2;
-    let mut start2 = end2;
-    while start2 > 0 {
-        let b = bytes[start2 - 1];
-        if b.is_ascii_alphanumeric() || b == b'_' {
-            start2 -= 1;
-        } else {
-            break;
-        }
-    }
-    &s[start2..end2] == "CORE"
-}
-
-/// Check if the match is a sigil-prefixed identifier ($print, @say, %exit, *dump)
-/// BUT NOT if it's a dereference call (&$print) or method call (->$print)
-fn is_sigil_prefixed_identifier(s: &str, op_start: usize) -> bool {
-    let bytes = s.as_bytes();
-    if op_start == 0 {
-        return false;
-    }
-
-    // Must be preceded by a sigil
-    if !matches!(bytes[op_start - 1], b'$' | b'@' | b'%' | b'*') {
-        return false;
-    }
-
-    // Security: If it's a sigil, we must ensure it's not being used in a way
-    // that triggers execution (like &$sub or ->$method).
-    // We scan backwards from the sigil (op_start - 1) skipping whitespace.
-    let mut i = op_start - 1;
-    while i > 0 && bytes[i - 1].is_ascii_whitespace() {
-        i -= 1;
-    }
-
-    if i > 0 {
-        let prev = bytes[i - 1];
-
-        // Block dereference execution (&$sub)
-        if prev == b'&' {
-            return false;
-        }
-
-        // Block method call (->$method)
-        if prev == b'>' && i > 1 && bytes[i - 2] == b'-' {
-            return false;
-        }
-
-        // Handle braced dereference &{ $sub }
-        if prev == b'{' {
-            i -= 1;
-            while i > 0 && bytes[i - 1].is_ascii_whitespace() {
-                i -= 1;
-            }
-            if i > 0 && bytes[i - 1] == b'&' {
-                return false;
-            }
-        }
-    }
-
-    true
-}
-
-/// Check if the match is a simple braced scalar variable ${print}
-/// Does NOT skip ${print()} or ${print + 1}
-fn is_simple_braced_scalar_var(s: &str, op_start: usize, op_end: usize) -> bool {
-    let bytes = s.as_bytes();
-
-    // Scan left for `${` (allow whitespace between)
-    let mut i = op_start;
-    while i > 0 && bytes[i - 1].is_ascii_whitespace() {
-        i -= 1;
-    }
-    if i < 1 || bytes[i - 1] != b'{' {
-        return false;
-    }
-    i -= 1;
-    while i > 0 && bytes[i - 1].is_ascii_whitespace() {
-        i -= 1;
-    }
-    if i < 1 || bytes[i - 1] != b'$' {
-        return false;
-    }
-
-    // Scan right for `}` (allow whitespace between)
-    let mut j = op_end;
-    while j < bytes.len() && bytes[j].is_ascii_whitespace() {
-        j += 1;
-    }
-    j < bytes.len() && bytes[j] == b'}'
-}
-
-/// Check if the match is package-qualified (Foo::print) but not CORE::
-fn is_package_qualified_not_core(s: &str, op_start: usize) -> bool {
-    let bytes = s.as_bytes();
-    if op_start < 2 || bytes[op_start - 1] != b':' || bytes[op_start - 2] != b':' {
-        return false;
-    }
-    // It's qualified, but we need to check it's not CORE::
-    !is_core_qualified(s, op_start)
-}
-
-/// Validate that an expression is safe for evaluation (non-mutating)
-///
-/// AC10.2: Safe evaluation mode validates expressions don't have side effects
-///
-/// This function uses a pre-compiled regex for performance and includes
-/// context-aware filtering to reduce false positives for:
-/// - Sigil-prefixed identifiers ($print, @say, %exit)
-/// - Simple braced scalar variables ${print}
-/// - Package-qualified names (Foo::print) unless CORE::
-/// - Single-quoted string literals ('print')
-///
-/// Note: Method calls ($obj->print) are intentionally NOT exempted because
-/// dangerous operations remain dangerous regardless of invocation syntax.
-fn validate_safe_expression(expression: &str) -> Option<String> {
-    // Check for assignment operators using regex to properly handle multi-char ops
-    // This avoids false positives for comparison operators (e.g., == contains =)
-    if let Some(re) = assignment_ops_re() {
-        for mat in re.find_iter(expression) {
-            let op = mat.as_str();
-            let start = mat.start();
-
-            // Allow harmless occurrences in single-quoted literals
-            if is_in_single_quotes(expression, start) {
-                continue;
-            }
-
-            // Check if it's strictly an assignment operator
-            match op {
-                "=" | "+=" | "-=" | "*=" | "/=" | "%=" | "**=" | ".=" | "&=" | "|=" | "^="
-                | "<<=" | ">>=" | "&&=" | "||=" | "//=" | "x=" => {
-                    return Some(format!(
-                        "Safe evaluation mode: assignment operator '{}' not allowed (use allowSideEffects: true)",
-                        op
-                    ));
-                }
-                _ => {}
-            }
-        }
-    }
-
-    // Check for dynamic subroutine calls &{...}
-    // This blocks tricks like &{"sys"."tem"}("ls")
-    if let Some(re) = deref_re() {
-        if re.is_match(expression) {
-            return Some(
-                "Safe evaluation mode: dynamic subroutine calls (&{...}) not allowed (use allowSideEffects: true)"
-                    .to_string(),
-            );
-        }
-    }
-
-    // Check for glob operations <*...> (anywhere in expression)
-    // This blocks filesystem access via globs
-    if let Some(re) = glob_re() {
-        if re.is_match(expression) {
-            return Some(
-                "Safe evaluation mode: glob operations (<*...>) not allowed (use allowSideEffects: true)"
-                    .to_string(),
-            );
-        }
-    }
-
-    // Check for file handle reads <$fh> or globs at start of expression
-    // This blocks state changes via reads like <STDIN> or <$fh>
-    if expression.trim().starts_with('<') {
-        return Some(
-            "Safe evaluation mode: file handle reads (<...>) and globs not allowed (use allowSideEffects: true)"
-                .to_string(),
-        );
-    }
-
-    // Check for mutating operations using pre-compiled regex
-    if let Some(re) = dangerous_ops_re() {
-        for mat in re.find_iter(expression) {
-            let op = mat.as_str();
-            let start = mat.start();
-            let end = mat.end();
-
-            // Allow harmless occurrences in single-quoted literals
-            if is_in_single_quotes(expression, start) {
-                continue;
-            }
-
-            // Allow sigil-prefixed identifiers ($print, @say, %exit, *printf)
-            if is_sigil_prefixed_identifier(expression, start) {
-                continue;
-            }
-
-            // Allow ${print} (simple scalar braced variable form)
-            if is_simple_braced_scalar_var(expression, start, end) {
-                continue;
-            }
-
-            // Allow package-qualified names unless it's CORE::
-            if is_package_qualified_not_core(expression, start) {
-                continue;
-            }
-
-            // Block: either bare op or CORE:: qualified
-            return Some(format!(
-                "Safe evaluation mode: potentially mutating operation '{}' not allowed (use allowSideEffects: true)",
-                op
-            ));
-        }
-    }
-
-    // Check for regex mutation operators (s///, tr///, y///)
-    // Handled separately to avoid false positives with escape sequences like \s in /\s+/
-    if let Some(re) = regex_mutation_re() {
-        if let Some(mat) = re.find(expression) {
-            let op = mat.as_str();
-            let start = mat.start();
-
-            // Allow sigil-prefixed identifiers ($s, $tr, $y)
-            if is_sigil_prefixed_identifier(expression, start) {
-                // It's a variable, allow it
-            } else if is_escape_sequence(expression, start) {
-                // It's an escape sequence like \s or \y, allow it
-            } else {
-                return Some(format!(
-                    "Safe evaluation mode: regex mutation operator '{}' not allowed (use allowSideEffects: true)",
-                    op.trim()
-                ));
-            }
-        }
-    }
-
-    // Check for increment/decrement operators
-    if expression.contains("++") || expression.contains("--") {
-        return Some(
-            "Safe evaluation mode: increment/decrement operators not allowed (use allowSideEffects: true)"
-                .to_string(),
-        );
-    }
-
-    // Check for backticks (shell execution)
-    if expression.contains('`') {
-        return Some(
-            "Safe evaluation mode: backticks (shell execution) not allowed (use allowSideEffects: true)"
-                .to_string(),
-        );
-    }
-
-    None
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2128,686 +1264,6 @@ mod tests {
             _ => return Err("Expected response".into()),
         }
         Ok(())
-    }
-
-    #[test]
-    fn test_parse_scope_variables_from_recent_output() -> Result<(), Box<dyn std::error::Error>> {
-        let adapter = DebugAdapter::new();
-        {
-            let mut output =
-                lock_or_recover(&adapter.recent_output, "test_parse_scope_variables.recent_output");
-            output.push_back("$foo = 42".to_string());
-            output.push_back("@arr = (1, 2, 3)".to_string());
-            output.push_back("%hash = {a => 1}".to_string());
-        }
-
-        let (vars, child_cache) = adapter.parse_scope_variables_from_output(11, 0, 20);
-        let names: Vec<&str> = vars.iter().map(|v| v.name.as_str()).collect();
-        assert!(names.contains(&"$foo"));
-        assert!(names.contains(&"@arr"));
-        assert!(names.contains(&"%hash"));
-        assert!(!child_cache.is_empty(), "expected child cache entries for expandable values");
-        Ok(())
-    }
-
-    #[test]
-    fn test_parse_scope_variables_are_sorted_for_stability()
-    -> Result<(), Box<dyn std::error::Error>> {
-        let lines = vec!["$zeta = 1".to_string(), "$alpha = 2".to_string(), "$mid = 3".to_string()];
-
-        let (vars, _child_cache) =
-            DebugAdapter::parse_scope_variables_from_lines(&lines, 11, 0, 20);
-        let names = vars.iter().map(|v| v.name.as_str()).collect::<Vec<_>>();
-        assert_eq!(names, vec!["$alpha", "$mid", "$zeta"]);
-        Ok(())
-    }
-
-    #[test]
-    fn test_capture_framed_debugger_output_isolated_by_marker()
-    -> Result<(), Box<dyn std::error::Error>> {
-        let adapter = DebugAdapter::new();
-        {
-            let mut output = lock_or_recover(
-                &adapter.recent_output,
-                "test_capture_framed_debugger_output.recent_output",
-            );
-            output.push_back("noise".to_string());
-            output.push_back(r#""DAP_BEGIN_100""#.to_string());
-            output.push_back("$a = 1".to_string());
-            output.push_back(r#""DAP_END_100""#.to_string());
-            output.push_back(r#""DAP_BEGIN_200""#.to_string());
-            output.push_back("$b = 2".to_string());
-            output.push_back(r#""DAP_END_200""#.to_string());
-        }
-
-        let lines = adapter
-            .capture_framed_debugger_output("DAP_BEGIN_200", "DAP_END_200", 200)
-            .ok_or("expected framed output for marker 200")?;
-        assert_eq!(lines, vec!["$b = 2".to_string()]);
-        Ok(())
-    }
-
-    #[test]
-    fn test_stack_trace_uses_recent_output_when_available() -> Result<(), Box<dyn std::error::Error>>
-    {
-        let mut adapter = DebugAdapter::new();
-        {
-            let mut output = lock_or_recover(
-                &adapter.recent_output,
-                "test_stack_trace_recent_output.recent_output",
-            );
-            output.push_back("# 0 main::compute at /tmp/script.pl line 20".to_string());
-            output.push_back("# 1 Foo::process called at /tmp/Foo.pm line 15".to_string());
-        }
-
-        let response = adapter.handle_request(1, "stackTrace", Some(json!({"threadId": 1})));
-        match response {
-            DapMessage::Response { success, body, .. } => {
-                assert!(success);
-                let body = body.ok_or("missing stackTrace body")?;
-                let frames = body
-                    .get("stackFrames")
-                    .and_then(|v| v.as_array())
-                    .ok_or("missing stackFrames")?;
-                assert!(
-                    frames.len() >= 2,
-                    "expected parsed frames from recent output, got {}",
-                    frames.len()
-                );
-            }
-            _ => return Err("expected stackTrace response".into()),
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_parse_evaluate_result_from_recent_output() -> Result<(), Box<dyn std::error::Error>> {
-        let adapter = DebugAdapter::new();
-        {
-            let mut output =
-                lock_or_recover(&adapter.recent_output, "test_parse_evaluate_result.recent_output");
-            output.push_back("$result = 123".to_string());
-        }
-
-        let parsed = adapter.parse_evaluate_result_from_output("$result");
-        let (value, ty) = parsed.ok_or("expected parsed evaluate result")?;
-        assert_eq!(value, "123");
-        assert_eq!(ty, "SCALAR");
-        Ok(())
-    }
-
-    // Tests for safe_eval false-positive filtering
-    #[test]
-    fn safe_eval_allows_identifiers_named_like_ops() {
-        // These should NOT be blocked - they're identifiers, not builtins
-        let allowed = [
-            "$print",           // scalar variable
-            "@say",             // array variable
-            "%exit",            // hash variable
-            "*printf",          // glob
-            "${print}",         // braced scalar variable
-            "${ print }",       // braced with spaces
-            "'print'",          // single-quoted string
-            "Foo::print",       // package-qualified
-            "My::Module::exit", // deeply qualified
-        ];
-
-        for expr in allowed {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_none(), "unexpected block for {expr:?}: {err:?}");
-        }
-    }
-
-    #[test]
-    fn safe_eval_still_blocks_real_ops() {
-        // These MUST be blocked - they're actual dangerous operations
-        let blocked = [
-            "print",
-            "print $x",
-            "say 'hello'",
-            "exit",
-            "exit 0",
-            "eval '$x'",
-            "eval { }",
-            "system 'ls'",
-            "exec '/bin/sh'",
-            "fork",
-            "kill 9, $$",
-            "CORE::print $x",
-            "CORE::GLOBAL::exit",
-            "$obj->print",
-            "$obj->system('ls')",
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    #[test]
-    fn test_safe_eval_mutating_regex_ops() {
-        let blocked = [
-            "$x =~ s/a/b/",
-            "s/a/b/",
-            "$x =~ tr/a/b/",
-            "tr/a/b/",
-            "y/a/b/",
-            "$x =~ y/a/b/", // Bound y/// form
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    #[test]
-    fn test_safe_eval_allows_regex_literals_with_escape_sequences() {
-        // These should NOT be blocked - they're regex patterns or identifiers, not mutations
-        // Note: Patterns using =~ are blocked by the assignment check (pre-existing behavior)
-        // so we test patterns without =~ here
-        let allowed = [
-            r#"/\s+/"#,    // \s in regex literal (no binding operator)
-            r#"/string/"#, // match containing 's'
-            r#"/tricky/"#, // match containing 'tr'
-            r#"/yay/"#,    // match containing 'y'
-            r#"$s"#,       // variable named $s
-            r#"$tr"#,      // variable named $tr
-            r#"$y"#,       // variable named $y
-            r#"qr/\s+/"#,  // compiled regex with \s
-        ];
-
-        for expr in allowed {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_none(), "unexpected block for {expr:?}: {err:?}");
-        }
-    }
-
-    #[test]
-    fn safe_eval_blocks_new_dangerous_ops() {
-        // Verify the extended deny-list works
-        let blocked = [
-            "eval '$code'",
-            "kill 9, $pid",
-            "exit 1",
-            "dump",
-            "fork",
-            "chroot '/tmp'",
-            "print STDERR 'x'",
-            "say 'hello'",
-            "printf '%s', $x",
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    #[test]
-    fn safe_eval_blocks_extended_ops_v2() {
-        // Verify the even more extended deny-list works (glob, readline, IPC, etc.)
-        let blocked = [
-            "glob '*'",
-            "readline $fh",
-            "ioctl $fh, 1, 1",
-            "srand",
-            "dbmopen %h, 'file', 0666",
-            "shmget $key, 10, 0666",
-            "select $r, $w, $e, 0",
-            "shutdown $socket, 2",
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    #[test]
-    fn safe_eval_blocks_mutation_and_resource_ops() {
-        // Verify newly added mutation and resource management operations are blocked
-        let blocked = [
-            "bless $ref, 'Class'",
-            "reset 'a-z'",
-            "umask 0022",
-            "binmode $fh",
-            "opendir $dh, '.'",
-            "closedir $dh",
-            "seek $fh, 0, 0",
-            "sysseek $fh, 0, 0",
-            "setpgrp",
-            "setpriority 0, 0, 10",
-            "formline",
-            "write",
-            "lock $ref",
-            "pipe $r, $w",
-            "socketpair $r, $w, 1, 1, 1",
-            "setsockopt $s, 1, 1, 1",
-            "utime 1, 1, 'file'",
-            "readdir $dh",
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    #[test]
-    fn test_safe_eval_blocks_dereference_execution() {
-        // These are variables (safe to access)
-        let allowed = ["$system", "@exec", "%fork"];
-
-        for expr in allowed {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_none(), "unexpected block for {expr:?}: {err:?}");
-        }
-
-        // These are dereference calls (NOT safe)
-        // &$system calls the sub ref in $system
-        // ->$system calls the method named in $system
-        let blocked = [
-            "&$system",
-            "& $system",
-            "&{$system}", // Braced form
-            "$obj->$system",
-            "$obj-> $system",
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
-    }
-
-    /// Helper to create a test stack frame
-    fn make_test_frame(id: i32, name: &str, path: &str, line: i32) -> StackFrame {
-        StackFrame {
-            id,
-            name: name.to_string(),
-            source: Source {
-                name: Some(path.split('/').next_back().unwrap_or(path).to_string()),
-                path: path.to_string(),
-                source_reference: None,
-            },
-            line,
-            column: 1,
-            end_line: None,
-            end_column: None,
-        }
-    }
-
-    /// Test helper: Filter frames using the same logic as handle_stack_trace (AC8.2.1)
-    fn filter_internal_frames(frames: Vec<StackFrame>) -> Vec<StackFrame> {
-        DebugAdapter::filter_user_visible_frames(frames)
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_removes_db_frames() {
-        // AC8.2.1: Filter internal frames from user-visible stack
-        let frames = vec![
-            make_test_frame(1, "main::hello", "/app/hello.pl", 10),
-            make_test_frame(2, "DB::DB", "/usr/share/perl/5.34/perl5db.pl", 100),
-            make_test_frame(3, "Foo::bar", "/app/lib/Foo.pm", 25),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert_eq!(filtered.len(), 2);
-        assert_eq!(filtered[0].name, "main::hello");
-        assert_eq!(filtered[1].name, "Foo::bar");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_removes_shim_frames() {
-        // AC8.2.1: Filter Devel::TSPerlDAP:: shim infrastructure frames
-        let frames = vec![
-            make_test_frame(1, "Devel::TSPerlDAP::init", "/shim/TSPerlDAP.pm", 50),
-            make_test_frame(2, "main::run", "/app/script.pl", 5),
-            make_test_frame(3, "Devel::TSPerlDAP::handle_break", "/shim/TSPerlDAP.pm", 200),
-            make_test_frame(4, "Utils::process", "/app/lib/Utils.pm", 42),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert_eq!(filtered.len(), 2);
-        assert_eq!(filtered[0].name, "main::run");
-        assert_eq!(filtered[1].name, "Utils::process");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_removes_perl5db_source() {
-        // AC8.2.1: Filter frames from perl5db.pl even with different names
-        let frames = vec![
-            make_test_frame(1, "main::start", "/app/main.pl", 1),
-            make_test_frame(2, "some_internal", "/usr/lib/perl5/perl5db.pl", 999),
-            make_test_frame(3, "App::process", "/app/lib/App.pm", 100),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert_eq!(filtered.len(), 2);
-        assert_eq!(filtered[0].name, "main::start");
-        assert_eq!(filtered[1].name, "App::process");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_mixed_internal_frames() {
-        // AC8.2.1: Comprehensive test with all types of internal frames mixed
-        let frames = vec![
-            // User frame 1
-            make_test_frame(1, "main::hello", "/app/hello.pl", 10),
-            // DB:: frame (internal)
-            make_test_frame(2, "DB::sub", "/usr/share/perl/5.34/perl5db.pl", 2000),
-            // User frame 2
-            make_test_frame(3, "Foo::bar", "/app/lib/Foo.pm", 25),
-            // Shim frame (internal)
-            make_test_frame(4, "Devel::TSPerlDAP::step", "/shim/TSPerlDAP.pm", 150),
-            // DB:: frame without perl5db.pl path (still filtered)
-            make_test_frame(5, "DB::breakpoint", "/some/other/path.pm", 50),
-            // User frame 3
-            make_test_frame(6, "Baz::qux", "/app/lib/Baz.pm", 75),
-            // perl5db.pl source frame (internal)
-            make_test_frame(7, "custom_handler", "/usr/lib/perl5/perl5db.pl", 1500),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        // Should only have user frames: main::hello, Foo::bar, Baz::qux
-        assert_eq!(filtered.len(), 3, "Expected 3 user frames, got {}", filtered.len());
-        assert_eq!(filtered[0].name, "main::hello");
-        assert_eq!(filtered[1].name, "Foo::bar");
-        assert_eq!(filtered[2].name, "Baz::qux");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_preserves_order() {
-        // Verify frame order is preserved after filtering
-        let frames = vec![
-            make_test_frame(1, "A::first", "/a.pm", 1),
-            make_test_frame(2, "DB::internal", "/perl5db.pl", 100),
-            make_test_frame(3, "B::second", "/b.pm", 2),
-            make_test_frame(4, "Devel::TSPerlDAP::shim", "/shim.pm", 50),
-            make_test_frame(5, "C::third", "/c.pm", 3),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert_eq!(filtered.len(), 3);
-        assert_eq!(filtered[0].name, "A::first");
-        assert_eq!(filtered[1].name, "B::second");
-        assert_eq!(filtered[2].name, "C::third");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_all_internal() {
-        // Edge case: all frames are internal
-        let frames = vec![
-            make_test_frame(1, "DB::main", "/perl5db.pl", 1),
-            make_test_frame(2, "Devel::TSPerlDAP::init", "/shim.pm", 10),
-            make_test_frame(3, "DB::sub", "/perl5db.pl", 50),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert!(filtered.is_empty(), "Expected empty stack after filtering all internal frames");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_no_internal() {
-        // Edge case: no internal frames to filter
-        let frames = vec![
-            make_test_frame(1, "main::start", "/app/main.pl", 1),
-            make_test_frame(2, "Lib::helper", "/app/lib/Lib.pm", 50),
-            make_test_frame(3, "Utils::format", "/app/lib/Utils.pm", 100),
-        ];
-
-        let filtered = filter_internal_frames(frames);
-
-        assert_eq!(filtered.len(), 3);
-        assert_eq!(filtered[0].name, "main::start");
-        assert_eq!(filtered[1].name, "Lib::helper");
-        assert_eq!(filtered[2].name, "Utils::format");
-    }
-
-    #[test]
-    fn test_stack_frame_filtering_empty_input() {
-        // Edge case: empty frame list
-        let frames: Vec<StackFrame> = vec![];
-        let filtered = filter_internal_frames(frames);
-        assert!(filtered.is_empty());
-    }
-
-    // AC8.2.4: Stack trace parsing tests for simple call chains (A → B → C)
-    #[test]
-    fn test_parse_stack_trace_simple_call_chain() {
-        let output = r#"# 0 main::compute_sum at /app/script.pl line 20
-# 1 Foo::process called at /app/lib/Foo.pm line 15
-# 2 main::start at /app/script.pl line 5"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-
-        assert_eq!(frames.len(), 3);
-
-        // Frame 0: main::compute_sum
-        assert_eq!(frames[0].id, 1);
-        assert_eq!(frames[0].name, "main::compute_sum");
-        assert_eq!(frames[0].source.path, "/app/script.pl");
-        assert_eq!(frames[0].line, 20);
-        assert_eq!(frames[0].source.name, Some("script.pl".to_string()));
-
-        // Frame 1: Foo::process
-        assert_eq!(frames[1].id, 2);
-        assert_eq!(frames[1].name, "Foo::process");
-        assert_eq!(frames[1].source.path, "/app/lib/Foo.pm");
-        assert_eq!(frames[1].line, 15);
-
-        // Frame 2: main::start
-        assert_eq!(frames[2].id, 3);
-        assert_eq!(frames[2].name, "main::start");
-        assert_eq!(frames[2].source.path, "/app/script.pl");
-        assert_eq!(frames[2].line, 5);
-    }
-
-    // AC8.2.4: Stack trace parsing for multi-file call stacks across packages
-    #[test]
-    fn test_parse_stack_trace_multi_file_packages() {
-        let output = r#"# 0 Utils::Helper::validate at /app/lib/Utils/Helper.pm line 42
-# 1 Data::Processor::transform called at /app/lib/Data/Processor.pm line 120
-# 2 Controller::API::handle_request at /app/controller/API.pm line 78
-# 3 main::dispatch called at /app/app.pl line 10"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-
-        assert_eq!(frames.len(), 4);
-        assert_eq!(frames[0].name, "Utils::Helper::validate");
-        assert_eq!(frames[1].name, "Data::Processor::transform");
-        assert_eq!(frames[2].name, "Controller::API::handle_request");
-        assert_eq!(frames[3].name, "main::dispatch");
-
-        // Verify cross-file navigation info is present
-        assert!(frames[0].source.path.contains("Utils/Helper.pm"));
-        assert!(frames[1].source.path.contains("Data/Processor.pm"));
-        assert!(frames[2].source.path.contains("controller/API.pm"));
-        assert!(frames[3].source.path.contains("app.pl"));
-    }
-
-    // AC8.2.4: Stack trace parsing for recursive calls with depth
-    #[test]
-    fn test_parse_stack_trace_recursive_calls() {
-        let output = r#"# 0 main::factorial at /app/math.pl line 5
-# 1 main::factorial called at /app/math.pl line 6
-# 2 main::factorial called at /app/math.pl line 6
-# 3 main::factorial called at /app/math.pl line 6
-# 4 main::compute at /app/math.pl line 10"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-
-        assert_eq!(frames.len(), 5);
-
-        // Verify recursive frames are all parsed correctly
-        assert_eq!(frames[0].name, "main::factorial");
-        assert_eq!(frames[1].name, "main::factorial");
-        assert_eq!(frames[2].name, "main::factorial");
-        assert_eq!(frames[3].name, "main::factorial");
-        assert_eq!(frames[4].name, "main::compute");
-
-        // Verify frame IDs are sequential
-        assert_eq!(frames[0].id, 1);
-        assert_eq!(frames[1].id, 2);
-        assert_eq!(frames[2].id, 3);
-        assert_eq!(frames[3].id, 4);
-        assert_eq!(frames[4].id, 5);
-    }
-
-    // AC8.2.4: Stack trace parsing for anonymous subroutines
-    #[test]
-    fn test_parse_stack_trace_anonymous_subs() {
-        let output = r#"# 0 main::__ANON__ at /app/callback.pl line 15
-# 1 Utils::map called at /app/lib/Utils.pm line 42
-# 2 main::process_items at /app/callback.pl line 10"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-
-        assert_eq!(frames.len(), 3);
-
-        // Verify anonymous sub is parsed (Perl uses __ANON__ for anonymous subs)
-        assert_eq!(frames[0].name, "main::__ANON__");
-        assert_eq!(frames[1].name, "Utils::map");
-        assert_eq!(frames[2].name, "main::process_items");
-    }
-
-    // AC8.2: Stack trace parsing with Windows paths
-    #[test]
-    fn test_parse_stack_trace_windows_paths() {
-        let output = r#"# 0 main::test at C:\workspace\script.pl line 10
-# 1 Foo::bar called at C:\workspace\lib\Foo.pm line 25"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-
-        assert_eq!(frames.len(), 2);
-        assert_eq!(frames[0].source.path, r"C:\workspace\script.pl");
-        assert_eq!(frames[0].source.name, Some("script.pl".to_string()));
-        assert_eq!(frames[1].source.path, r"C:\workspace\lib\Foo.pm");
-        assert_eq!(frames[1].source.name, Some("Foo.pm".to_string()));
-    }
-
-    // AC8.2: Stack trace parsing with empty output
-    #[test]
-    fn test_parse_stack_trace_empty_output() {
-        let output = "";
-        let frames = DebugAdapter::parse_stack_trace(output);
-        assert!(frames.is_empty());
-    }
-
-    // AC8.2: Stack trace parsing with malformed output
-    #[test]
-    fn test_parse_stack_trace_malformed_output() {
-        let output = r#"Random output that doesn't match
-Some error message
-DB<1>"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-        assert!(frames.is_empty());
-    }
-
-    // AC8.2.1: Integration test - parse and filter combined
-    #[test]
-    fn test_parse_and_filter_stack_trace() {
-        let output = r#"# 0 main::user_func at /app/script.pl line 10
-# 1 DB::DB called at /usr/share/perl/5.34/perl5db.pl line 100
-# 2 Foo::process at /app/lib/Foo.pm line 25
-# 3 Devel::TSPerlDAP::handle_break called at /shim/TSPerlDAP.pm line 50
-# 4 main::start at /app/script.pl line 5"#;
-
-        let frames = DebugAdapter::parse_stack_trace(output);
-        assert_eq!(frames.len(), 5);
-
-        // Apply filtering
-        let filtered = filter_internal_frames(frames);
-
-        // Should only have user frames after filtering
-        assert_eq!(filtered.len(), 3);
-        assert_eq!(filtered[0].name, "main::user_func");
-        assert_eq!(filtered[1].name, "Foo::process");
-        assert_eq!(filtered[2].name, "main::start");
-    }
-
-    #[test]
-    fn test_safe_eval_bypass_prevention() {
-        // These patterns attempt to bypass safe evaluation checks
-        let bypasses = [
-            "&{'sys'.'tem'}('ls')", // Dynamic function name via concatenation
-            "& { 'sys' . 'tem' }",  // Dynamic function name with spaces
-            "<*.txt>",              // Glob operator for filesystem access
-            "CORE::print",          // Explicitly blocked by dangerous ops regex
-        ];
-
-        for expr in bypasses {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "Expression '{}' should be blocked but was allowed", expr);
-        }
-    }
-
-    #[test]
-    fn test_safe_eval_assignment_ops_precision() {
-        // These are comparison/binding operators (SAFE) but were previously blocked
-        // because they contain '='
-        let allowed = [
-            "$a == $b",
-            "$a != $b",
-            "$a <= $b",
-            "$a >= $b",
-            "$a <=> $b",
-            "$a =~ /regex/",
-            "$a !~ /regex/",
-            "$a ~~ $b", // Smart match
-            // Logical ops
-            "$a && $b",
-            "$a || $b",
-            "$a // $b",
-            // Bitwise ops
-            "$a & $b",
-            "$a | $b",
-            "$a ^ $b",
-            "$a << $b",
-            "$a >> $b",
-            // Range
-            "1..10",
-        ];
-
-        for expr in allowed {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_none(), "unexpected block for {expr:?}: {err:?}");
-        }
-
-        // These are strict assignment operators (UNSAFE) and MUST be blocked
-        let blocked = [
-            "$a = 1",
-            "$a += 1",
-            "$a -= 1",
-            "$a *= 1",
-            "$a /= 1",
-            "$a %= 1",
-            "$a **= 1",
-            "$a .= 's'",
-            "$a &= 1",
-            "$a |= 1",
-            "$a ^= 1",
-            "$a <<= 1",
-            "$a >>= 1",
-            "$a &&= 1",
-            "$a ||= 1",
-            "$a //= 1",
-            "$a x= 3", // Repetition assignment
-        ];
-
-        for expr in blocked {
-            let err = validate_safe_expression(expr);
-            assert!(err.is_some(), "expected block for {expr:?}");
-        }
     }
 
     #[test]
